requirements.txt
psycopg2-binary
colorama
Flask==2.3.2